moved course to courselist
